Create wrapper function to access collision arrays
diffwrf reports difference; need to check.

Previously, we pass functions to access a specific collision array, e.g.
get_cwll into the collision routines. This is apparently not allowed
when we declare these routines as omp targets. Thus, we now have to pass
in an integer to a wrapper function `get_cw_generic` to choose the array to access.

Because this requires the collision routines (defined in a separate
module) to access the wrapper function (defined in the fast_sbm module),
a cycle is created. Therefore we get rid of the collision module and
move all the collision routines inside the fast_sbm module, making the
commit very large (need to fix).