Create wrapper for accessing collision entries
Previously, we pass functions to access a specific collision entry, e.g.
get_cwll into the collision routines. This is apparently not allowed
when we declare these routines as omp targets. Thus, we now have to pass
in an integer to a wrapper function `get_cw_wrapper` to choose the entry to access.

We also declare global variables used in the collision routines as
device targets, and hide their I/O calls based on the macro OFFLOAD.